requirements.txt
streamlit
pandas
altair
numpy
sqlalchemy
psycopg2-binary
bcrypt
openai
python-dotenv
